category create is working

+++ app/Http/Controllers/AdminUsersController.php
@@ -135,9 +135,12 @@ class AdminUsersController extends Controller
 
     {
 
-       $user = User::findOrFail($id);
-       unlink(public_path().$user->photo->file);
-       $user->delete();
+        $user = User::findOrFail($id);
+        if ($user->photo) {
+
+            unlink(public_path() . $user->photo->file);
+        }
+        $user->delete();
 
         Session::flash('deleted_user', 'The user has been deleted');
